requirements.txt
xformers==0.0.16
pytorch_lightning==1.4.2
einops
open-clip-torch
omegaconf
torchmetrics==0.6.0
triton==2.0.0
opencv-python-headless
scipy
matplotlib
lpips
gradio
chardet
transformers
facexlib
